Remove versioning in graphql-federated-graph
Only v4 left. Not renaming it for now to keep a clean diff. First PR in a batch of simplifications made possible by the removal of managed graphs.

closes GB-7947
