+++ engine/crates/federated-graph/src/federated_graph.rs
@@ -1,6 +1,3 @@
-mod v1;
-mod v2;
-mod v3;
 mod v4;
 
-pub use self::{v1::FederatedGraphV1, v2::FederatedGraphV2, v3::FederatedGraphV3, v4::*};
+pub use self::v4::*;
