+++ engine/crates/federated-graph/src/federated_graph/v4/ids.rs
@@ -1,10 +1,10 @@
-use super::FederatedGraph;
+use super::*;
 
 macro_rules! id_newtypes {
     ($($storage:ident [ $name:ident ] -> $out:ident,)*) => {
         $(
             #[derive(Debug, Clone, Copy, PartialOrd, Ord, PartialEq, Eq, Hash, serde::Serialize, serde::Deserialize)]
-            pub struct $name(usize);
+            pub struct $name(pub usize);
 
             impl From<$name> for usize {
               fn from(value: $name) -> usize {
@@ -19,15 +19,15 @@ macro_rules! id_newtypes {
             }
 
             impl std::ops::Index<$name> for FederatedGraph {
-                type Output = super::$out;
+                type Output = $out;
 
-                fn index(&self, index: $name) -> &super::$out {
+                fn index(&self, index: $name) -> &$out {
                     &self.$storage[index.0]
                 }
             }
 
             impl std::ops::IndexMut<$name> for FederatedGraph {
-                fn index_mut(&mut self, index: $name) -> &mut super::$out {
+                fn index_mut(&mut self, index: $name) -> &mut $out {
                     &mut self.$storage[index.0]
                 }
             }
@@ -36,6 +36,16 @@ macro_rules! id_newtypes {
 }
 
 id_newtypes! {
-    type_definitions[TypeDefinitionId] -> TypeDefinitionRecord,
+    authorized_directives[AuthorizedDirectiveId] -> AuthorizedDirective,
+    directives[DirectiveId] -> Directive,
     enum_values[EnumValueId] -> EnumValueRecord,
+    fields[FieldId] -> Field,
+    input_objects[InputObjectId] -> InputObject,
+    input_value_definitions[InputValueDefinitionId] -> InputValueDefinition,
+    interfaces[InterfaceId] -> Interface,
+    objects[ObjectId] -> Object,
+    strings[StringId] -> String,
+    subgraphs[SubgraphId] -> Subgraph,
+    type_definitions[TypeDefinitionId] -> TypeDefinitionRecord,
+    unions[UnionId] -> Union,
 }
