+++ engine/crates/federated-graph/src/federated_graph/v4/root_operation_types.rs
@@ -1,0 +1,8 @@
+use crate::ObjectId;
+
+#[derive(Clone)]
+pub struct RootOperationTypes {
+    pub query: ObjectId,
+    pub mutation: Option<ObjectId>,
+    pub subscription: Option<ObjectId>,
+}
