+++ engine/crates/federated-graph/src/lib.rs
@@ -18,9 +18,6 @@ pub use from_sdl::{from_sdl, DomainError};
 
 #[derive(serde::Serialize, serde::Deserialize, Clone, Debug)]
 pub enum VersionedFederatedGraph {
-    V1(FederatedGraphV1),
-    V2(FederatedGraphV2),
-    V3(FederatedGraphV3),
     Sdl(String),
 }
 
@@ -32,15 +29,11 @@ impl VersionedFederatedGraph {
     pub fn into_federated_sdl(self) -> Result<String, DomainError> {
         Ok(match self {
             VersionedFederatedGraph::Sdl(sdl) => sdl,
-            other => render_federated_sdl(&other.into_latest()?).unwrap(),
         })
     }
 
     pub fn into_latest(self) -> Result<FederatedGraph, DomainError> {
         Ok(match self {
-            VersionedFederatedGraph::V1(v1) => VersionedFederatedGraph::V2(FederatedGraphV2::from(v1)).into_latest()?,
-            VersionedFederatedGraph::V2(v2) => VersionedFederatedGraph::V3(FederatedGraphV3::from(v2)).into_latest()?,
-            VersionedFederatedGraph::V3(v3) => v3.into(),
             VersionedFederatedGraph::Sdl(sdl) => from_sdl(&sdl)?,
         })
     }
